# dist/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "My API",
        "description": "Description",
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [],
    "schemes": [
        "http"
    ],
    "consumes": [],
    "produces": [],
    "paths": {
        "/user": {
            "post": {
                "tags": [],
                "description": "",
                "parameters": [
                    {
                        "name": "obj",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/todo": {
            "post": {
                "tags": [],
                "description": "",
                "parameters": [],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/todo/{id}": {
            "put": {
                "tags": [],
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {}
            }
        },
        "/todos/{userId}": {
            "get": {
                "tags": [],
                "description": "",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {}
}